==> 1_Data_Curation.ipynb <==
import os
import sys
import numpy as np
import pandas as pd
from rdkit.Chem import PandasTools
from rdkit import Chem
import molvs as mv

#INITIAL TRAINING DATASET
combinedDF = pd.DataFrame(columns=['SMILES','ID'])

for filename in os.listdir(os.getcwd()):
    if filename.endswith('.smiles'):
        df = pd.read_csv(filename, header=None, sep='\t')
        df.columns=['SMILES','ID',filename[:-7]]
        combinedDF = combinedDF.merge(df, on=['SMILES', 'ID'], how='outer')

combinedDF.shape

combinedDF.head()

#df_lead = pd.read_csv('.../.../tox21_10k_challenge_test.smiles', sep='\t')
#df_lead.head()

#THE FINAL TEST SET
df_score_smiles = pd.read_csv('/.../.../tox21_10k_challenge_score.smiles', sep='\t')
df_score_results = pd.read_csv('/.../.../Tox21_Restart/tox21_10k_challenge_score.txt', sep='\t')

df_test = pd.concat([df_score_smiles, df_score_results]).groupby('Sample ID', as_index=False, sort=False).first().fillna('NA')

df_test = df_test.drop('Sample ID', axis=1).replace('x', np.nan)
df_test.rename(columns={'#SMILES': 'smiles'}, inplace=True)
df_test.shape

df_test.head()

#LEADERBOARD DATA SET
df_leaderbd = PandasTools.LoadSDF('/.../.../tox21_10k_challenge_test.sdf')

#Drop duplicates by ID
#df_train_lead = pd.concat([df_train, df_leaderbd])
#df_train_lead = df_train_lead.drop_duplicates(['ID'], keep = 'first')
#tox21_df.info()

#Select only useful columns: assays and the ligands(ROMol) total = 13
keep_cols = ['NR-AR', 'NR-AR-LBD', 'NR-AhR', 'NR-Aromatase', 'NR-ER', 'NR-ER-LBD', 'NR-PPAR-gamma',
            'SR-ARE', 'SR-ATAD5', 'SR-HSE', 'SR-MMP', 'SR-p53', 'ROMol']
df_leaderbd = df_leaderbd[keep_cols]

#Convert Mol to smiles
df_leaderbd['smiles'] = df_leaderbd['ROMol'].apply(Chem.MolToSmiles)
df_leaderbd = df_leaderbd.drop('ROMol', axis=1)

cols = df_leaderbd.columns.tolist()
cols = cols[-1:] + cols[:-1]
df_leaderbd = df_leaderbd[cols]

df_leaderbd.head()

df_leaderbd.shape

#STANDARDIZATION STEP
#Function to get parent of a smiles
def get_parent_smile(smile):
    try:
        st = mv.Standardizer()
        mol = Chem.MolFromSmiles(smile)
        smts = Chem.MolFromSmarts("[!#1&!#5&!#6&!#7&!#8&!#9&!#14&!#15&!#16&!#17&!#34&!#35&!#53]~*")
        if mol.HasSubstructMatch(smts) == False:
            mols = st.charge_parent(mol)
            return Chem.MolToSmiles(mols)
        else:
            #print(smile)
            return 'problematic'
    except:
        return 'NaN'
    
#Clean and standardize the data
def clean_data(data):
    #remove missing smiles
    data = data[~(data['smiles'].isnull())]
    
    #Standardize and get parent with molvs
    data["smiles_parent"] = data.smiles.apply(get_parent_smile)
    data = data[~(data['smiles_parent'] == "NaN")]
    
    #Identifiy compounds that fail SMILE conversion
    pCompounds = data[data['smiles_parent']=='problematic']['smiles']
    data = data[~(data['smiles_parent'] == 'problematic')]
    print(len(pCompounds))
    
    return data

#Generate Unique InchiKey for Identification purpose
def inchikey_gen(smile):
    try:
        m = Chem.MolFromSmiles(smile)
        m_ = Chem.MolToInchi(m)
        m__ = Chem.InchiToInchiKey(m_)
        return m__
    except:
        return 'Failed'

def remove_pc(data):
    #Remove compunds that failed SMILE Conversion
    pCompounds = data[data['smiles_parent']=='problematic']['smiles']
    data = data[~(data['smiles_parent'] == 'problematic')]
    print(len(pCompounds))
    return data

#df_train = combinedDF.rename(columns={'SMILES': 'smiles'})
df_train.columns = df_train.columns.str.lower()
df_train = clean_data(df_train)

df_train.shape

df_leaderbd = clean_data(df_leaderbd)

df_test = clean_data(df_test)

#Drop duplicates by majority rule
#If a compound is duplicated, record the most occuring activity, if the activities for the dupplicates occur same number
#of times e.g. 4 duplicates with 2 active and 2 inactives, such compound (all its duplicates) should be considered 
#ambiguous and removed. If no duplicate, leave the compound.

def moder(x):
    m = pd.Series.mode(x)
    if len(m) == 1: 
        return m


#assays = ['NR-AR', 'NR-AR-LBD', 'NR-AhR', 'NR-Aromatase', 'NR-ER', 'NR-ER-LBD', 'NR-PPAR-gamma', 'SR-ARE', 'SR-ATAD5', 'SR-HSE', 'SR-MMP', 'SR-p53']
assays = ['nr-ar', 'nr-ar-lbd', 'nr-ahr', 'nr-aromatase', 'nr-er', 'nr-er-lbd', 'nr-ppar-gamma', 'sr-are', 'sr-atad5', 'sr-hse', 'sr-mmp', 'sr-p53']
def dedup_majority_rule(data):
    dataDF = pd.DataFrame(columns=['smiles_parent'])
    for i in assays:
        a = data[['smiles_parent', i]]
        res = a.groupby('smiles_parent')[i].apply(moder).reset_index(level=1, drop=True).reset_index()
        dataDF = dataDF.merge(res, on=['smiles_parent'], how='outer')
    return dataDF

    #res.head()
#combinedDF

#testDF = df_test.rename(columns = {'smiles_parent': 's_p'})
df_test.columns = df_test.columns.str.lower()
testDF = dedup_majority_rule(df_test)
testDF.shape

df_leaderbd.columns = df_leaderbd.columns.str.lower()
leaderbdDF = dedup_majority_rule(df_leaderbd)
leaderbdDF.shape

trainDF = dedup_majority_rule(df_train)
trainDF.shape

#Concat trainDF and leaderbdDF the training data
#Retain testDF as test data set

#trainDF.to_csv('trainDF.csv')
