==> 2_FeatureGen.ipynb <==
import pandas as pd
import numpy as np

trainDF = pd.read_csv('/.../.../trainDF.csv', header=0).drop('Unnamed: 0', axis = 1)
trainDF.head()

testDF = pd.read_csv('/.../.../testDF.csv', header=0).drop('Unnamed: 0', axis = 1)
testDF.head()

testDF.values[:,0][0]

from rdkit import Chem
from rdkit.Chem import Descriptors
from rdkit import Chem, DataStructs
from rdkit.Chem import AllChem, MACCSkeys
from rdkit.ML.Descriptors import MoleculeDescriptors

nms = [ x[0] for x in Descriptors._descList ] #list of all RDKit Descriptors
calc = MoleculeDescriptors.MolecularDescriptorCalculator( nms )
keep_cols = ['NR-AR', 'NR-AR-LBD', 'NR-AhR', 'NR-Aromatase', 'NR-ER', 'NR-ER-LBD', 'NR-PPAR-gamma',
            'SR-ARE', 'SR-ATAD5', 'SR-HSE', 'SR-MMP', 'SR-p53']

def featurize(data):
    
    mols = []
    X_mdesc = []
    X_fp = []
    X_maccs = []
    y = [] 
    for i in range(len(data)):
        smi = data[:,0][i]
        outcome = data[:,1:13][i]

        smi = Chem.MolFromSmiles(smi)
        if type(smi) != type(None):
            desc = calc.CalcDescriptors(smi)
            
            maccs = MACCSkeys.GenMACCSKeys(smi)
            
            fp = AllChem.GetMorganFingerprintAsBitVect(smi, 2, nBits=1024)
            fp_vect = np.zeros((1,))
            DataStructs.ConvertToNumpyArray(fp, fp_vect)

            mols.append(smi)
            X_mdesc.append(desc)
            X_fp.append(fp_vect)
            X_maccs.append(maccs)
            y.append(outcome)
            
        else:
            print("Failed for ID: %s"%smi)
    #print("Details for %s "% desc)
    print("Imported smiles: %s"%len(data))
    print("Smiles converted to MolDesc: %s"%len(mols))
    X1 = pd.DataFrame(np.asarray(X_mdesc))#descriptors
    X2 = pd.DataFrame(np.asarray(X_fp))#Morgan FP
    X3 = pd.DataFrame(np.asarray(X_maccs))#MACCS FP
    
    y = pd.DataFrame(np.asarray(y), columns = keep_cols)
    X_ = pd.concat([X1, X2, X3, y], axis = 1)
    return X_

testDF_fp = featurize(testDF.values)
trainDF_fp = featurize(trainDF.values)

testDF_fp.to_csv('testDF_rdkit.csv', index=False)
trainDF_fp.to_csv('trainDF_rdkit.csv', index=False)
